Move `Composition` to its own module
